+++ Jakefile.js
@@ -48,6 +48,7 @@ task('build', ['clear'], function () {
             jsoneditorSrc + 'js/appendnode.js',
             jsoneditorSrc + 'js/contextmenu.js',
             jsoneditorSrc + 'js/history.js',
+            jsoneditorSrc + 'js/modebox.js',
             jsoneditorSrc + 'js/searchbox.js',
             jsoneditorSrc + 'js/highlighter.js',
             jsoneditorSrc + 'js/util.js',
@@ -228,6 +229,7 @@ task('webapp', ['build', 'minify'], function () {
     jake.cpR(webAppSrc + 'robots.txt', webApp);
     jake.cpR(webAppSrc + 'datapolicy.txt', webApp);
     jake.cpR(webAppSrc + 'index.html', webApp);
+    jake.cpR(webAppSrc + 'chrome_app_counter.html', webApp);
     jake.cpR(webAppSrc + 'favicon.ico', webApp);
     jake.cpR(webAppSrc + 'fileretriever.php', webApp);
     jake.cpR(webAppSrc + 'googlea47c4a0b36d11021.html', webApp);