+++ examples/requirejs_demo/scripts/main.js
@@ -1,24 +1,24 @@
-var module = "../../../jsoneditor";
+var module = '../../../jsoneditor';
 require([module], function (jsoneditor) {
     // create the editor
-    var container = document.getElementById("jsoneditor");
+    var container = document.getElementById('jsoneditor');
     var editor = new jsoneditor.JSONEditor(container);
 
     // set json
-    document.getElementById("setJSON").onclick = function () {
+    document.getElementById('setJSON').onclick = function () {
         var json = {
-            "array": [1, 2, 3],
-            "boolean": true,
-            "null": null,
-            "number": 123,
-            "object": {"a": "b", "c": "d"},
-            "string": "Hello World"
+            'array': [1, 2, 3],
+            'boolean': true,
+            'null': null,
+            'number': 123,
+            'object': {'a': 'b', 'c': 'd'},
+            'string': 'Hello World'
         };
         editor.set(json);
     };
 
     // get json
-    document.getElementById("getJSON").onclick = function () {
+    document.getElementById('getJSON').onclick = function () {
         var json = editor.get();
         alert(JSON.stringify(json, null, 2));
     };