+++ jsoneditor/js/treeeditor.js
@@ -546,14 +546,9 @@ TreeEditor.prototype._createFrame = function () {
 
     // create undo/redo buttons
     if (this.history) {
-        // create separator
-        var separator = document.createElement('span');
-        separator.innerHTML = '&nbsp;';
-        this.menu.appendChild(separator);
-
         // create undo button
         var undo = document.createElement('button');
-        undo.className = 'undo';
+        undo.className = 'undo separator';
         undo.title = 'Undo last action (Ctrl+Z)';
         undo.onclick = function () {
             editor._onUndo();
@@ -577,6 +572,12 @@ TreeEditor.prototype._createFrame = function () {
             redo.disabled = !editor.history.canRedo();
         };
         this.history.onChange();
+    }
+
+    // create mode box
+    if (this.options && this.options.modes && this.options.modes.length) {
+        var modeBox = createModeBox(this, this.options.modes, this.options.mode);
+        this.menu.appendChild(modeBox);
     }
 
     // create search box