+++ HISTORY.md
@@ -3,6 +3,12 @@
 http://jsoneditoronline.org
 
 
+## not yet released, version 2.3.0
+
+- Implemented an option `modes`, which creates a menu in the editor
+  where the user can switch between the selected editor modes.
+
+
 ## 2013-08-01, version 2.2.2
 
 - Fixed non working option `indentation`.